feat(ts): add package and rename file
