+++ package.json
@@ -22,10 +22,12 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.10.0",
+    "@types/node": "^22.5.4",
     "eslint": "^9.10.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
     "globals": "^15.9.0",
-    "prettier": "^3.3.3"
+    "prettier": "^3.3.3",
+    "typescript": "^5.6.2"
   }
 }
