+++ src/services/bot.ts
@@ -1,3 +1,4 @@
+import { Client, Message } from "discord.js";
 import { checkStreams } from "../api/twitchApi.js";
 import { addCommand } from "../commands/add.js";
 import { deleteCommand } from "../commands/delete.js";
@@ -11,7 +12,7 @@ import { checkForReminders } from "./reminder.js";
 let streamStatus = {};
 let lastStreamTimestamps = {};
 
-export async function startBot(client) {
+export async function startBot(client: Client) {
   setInterval(() => {
     checkStreams(
       client,
@@ -23,7 +24,7 @@ export async function startBot(client) {
   }, config.twitchRefreshTime);
 }
 
-export async function handleMessage(client, msg) {
+export async function handleMessage(client: Client , msg: Message) {
   if (msg.author.bot) return;
   const stdin = msg.content.trim().split(" ");
   switch (stdin[0]) {
