+++ src/services/reminder.ts
@@ -1,5 +1,5 @@
 import { config } from "../config.js";
-import { sendReminder } from "./../messages/reminderStream.js";
+import { sendReminder } from "../messages/reminderStream.js";
 import { getStreamers } from "./database.js";
 const msInADay = 24 * 60 * 60 * 1000;
